Yay! TypeDissector changes are good
